+++ src/accessories/extras/extras.handler.js
@@ -38,7 +38,10 @@ class Handler {
   async change(context, accessory, subtype, historyService) {
     if (context.oldValue !== context.newValue) {
       if (!this.configured) {
-        logger.debug('Extras: Handler not configured yet. Skipping CHANGE event.');
+        logger.debug(
+          'Handler not configured yet. Skipping CHANGE event.',
+          `${accessory.displayName} (${accessory.context.config.subtype})`
+        );
         return;
       }
 
@@ -70,7 +73,10 @@ class Handler {
 
   async get(accessory, subtype, ownCharacteristic) {
     if (!this.configured) {
-      logger.debug('Extras: Handler not configured yet. Skipping GET event.');
+      logger.debug(
+        'Handler not configured yet. Skipping GET event.',
+        `${accessory.displayName} (${accessory.context.config.subtype})`
+      );
       return false;
     }
 
@@ -139,7 +145,10 @@ class Handler {
 
   async set(state, accessory, subtype, ownCharacteristic) {
     if (!this.configured) {
-      logger.debug('Extras: Handler not configured yet. Skipping SET event.');
+      logger.debug(
+        'Handler not configured yet. Skipping SET event.',
+        `${accessory.displayName} (${accessory.context.config.subtype})`
+      );
       return;
     }